requirements.txt
streamlit>=1.28.0
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pypdf>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
